Moved aocal into the vcstools module
